// src/app/enums/masterdatatype.enum.ts
export enum MasterDataType {
  vehicleModel = 1,
  vehicleType = 2,
  seatType = 3,
  country = 4,
  hostessSeat = 5,
  backRowSeat = 6,
  fastenerType = 7,
  portableFixed = 8,
  doorType = 9,
  armrest = 10,
  heating = 11,
  vibration = 12,
  mercedesStar = 13,
  electrical = 14,
  locking = 15,
  netType = 16
}